+++ NEWS.md
@@ -3,6 +3,7 @@
 
 * trailing_semicolon_linter (#147, @gaborcsardi)
 
+* Fixed lint_package bug where cache was not caching (#146, @schloerke)
 * Commas linter handles missing arguments calls properly (#145)
 
 # lintr 1.0.0 #